pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spec-runner"
version = "0.1.0"
description = "Executable spec runner for yaml contract-spec blocks in Markdown."
readme = "README.md"
requires-python = ">=3.10"
# Keep this as a string for compatibility with setuptools' PEP 621 validation.
license = "MIT"
license-files = ["LICENSE"]
authors = [{name = "jon"}]

dependencies = [
  "PyYAML>=6",
]

[project.urls]
Repository = "https://example.invalid/spec-runner"

[project.scripts]
spec-runner-conformance = "spec_runner.cli:conformance_runner_main"
spec-runner-parity = "spec_runner.cli:compare_parity_main"
spec-runner-validate-report = "spec_runner.spec_lang_commands:validate_report_main"

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "build>=1",
  "twine>=6",
  "ruff>=0.8",
  "mypy>=1.11",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["spec_runner*"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["F", "E9"]

[tool.mypy]
python_version = "3.10"
files = ["spec_runner"]
check_untyped_defs = false
warn_unused_ignores = true
warn_redundant_casts = true
disable_error_code = ["import-untyped"]
